Fix: Disable user scrolling in overview pager to not interfere with horizontal scrolling on filter chips
This commit disables manual swiping between pages in the `HorizontalPager` on the `OverviewScreen`.

The `userScrollEnabled` property has been set to `false`, meaning navigation between the "Filters and Table" view and the "Charts" view can now only be controlled programmatically, not by user gestures. This improves the predictability of the UI.